REFACTOR: Remove method to load a single concept.
Concepts can depend on each other various ways, e.g., morphisms and
subobject relations have domain and codomain objects. In order to safely
and correctly load a single concept, we should also load all dependent
concepts, possibly recursively. That's more trouble than it's worth. From
now on, concepts are an all-or-nothing affair: either they're loaded or
they're not.

If the ontology gets big enough, we may have to revisit this decision but
for the foreseeable future it should not be a problem.